QuantiFire IDE (V2) /QuantiFire ICP/requirements.txt
yfinance
pandas
beautifulsoup4
requests
openai
lxml
wcwidth
duckduckgo-search
apify-client
gspread
google-auth
google-auth-oauthlib
google-auth-httplib2

